corrected file names & skipped duplicated tests
