--- frontend/src/lib/components/Forms/ModelForm/ProjectForm.svelte
@@ -1,42 +1,0 @@
-<script lang="ts">
-	import AutocompleteSelect from '../AutocompleteSelect.svelte';
-	import TextField from '$lib/components/Forms/TextField.svelte';
-	import Select from '../Select.svelte';
-	import { getOptions } from '$lib/utils/crud';
-	import type { SuperValidated } from 'sveltekit-superforms';
-	import type { ModelInfo, CacheLock } from '$lib/utils/types';
-	import * as m from '$paraglide/messages.js';
-
-	export let form: SuperValidated<any>;
-	export let model: ModelInfo;
-	export let cacheLocks: Record<string, CacheLock> = {};
-	export let formDataCache: Record<string, any> = {};
-	export let initialData: Record<string, any> = {};
-</script>
-
-{#if model.urlModel === 'projects'}
-	<AutocompleteSelect
-		{form}
-		options={getOptions({ objects: model.foreignKeys['folder'] })}
-		field="folder"
-		cacheLock={cacheLocks['folder']}
-		bind:cachedValue={formDataCache['folder']}
-		label={m.domain()}
-		hidden={initialData.folder}
-	/>
-	<TextField
-		{form}
-		field="internal_reference"
-		label={m.internalReference()}
-		cacheLock={cacheLocks['internal_reference']}
-		bind:cachedValue={formDataCache['internal_reference']}
-	/>
-	<Select
-		{form}
-		options={model.selectOptions['lc_status']}
-		field="lc_status"
-		label={m.lcStatus()}
-		cacheLock={cacheLocks['lc_status']}
-		bind:cachedValue={formDataCache['lc_status']}
-	/>
-{/if}
