+++ frontend/src/lib/components/Forms/ModelForm/ReferenceControlsForm.svelte
@@ -15,51 +15,49 @@
 	export let initialData: Record<string, any> = {};
 </script>
 
-{#if model.urlModel === 'reference-controls'}
-	<TextField
-		{form}
-		field="ref_id"
-		label={m.ref()}
-		cacheLock={cacheLocks['ref_id']}
-		bind:cachedValue={formDataCache['ref_id']}
-	/>
-	<Select
-		{form}
-		options={model.selectOptions['category']}
-		field="category"
-		label={m.category()}
-		cacheLock={cacheLocks['category']}
-		bind:cachedValue={formDataCache['category']}
-	/>
-	<Select
-		{form}
-		options={model.selectOptions['csf_function']}
-		field="csf_function"
-		label={m.csfFunction()}
-		cacheLock={cacheLocks['csf_function']}
-		bind:cachedValue={formDataCache['csf_function']}
-	/>
-	<TextArea
-		{form}
-		field="annotation"
-		label={m.annotation()}
-		cacheLock={cacheLocks['annotation']}
-		bind:cachedValue={formDataCache['annotation']}
-	/>
-	<TextField
-		{form}
-		field="provider"
-		label={m.provider()}
-		cacheLock={cacheLocks['provider']}
-		bind:cachedValue={formDataCache['provider']}
-	/>
-	<AutocompleteSelect
-		{form}
-		options={getOptions({ objects: model.foreignKeys['folder'] })}
-		field="folder"
-		cacheLock={cacheLocks['folder']}
-		bind:cachedValue={formDataCache['folder']}
-		label={m.domain()}
-		hidden={initialData.folder}
-	/>
-{/if}
+<TextField
+	{form}
+	field="ref_id"
+	label={m.ref()}
+	cacheLock={cacheLocks['ref_id']}
+	bind:cachedValue={formDataCache['ref_id']}
+/>
+<Select
+	{form}
+	options={model.selectOptions['category']}
+	field="category"
+	label={m.category()}
+	cacheLock={cacheLocks['category']}
+	bind:cachedValue={formDataCache['category']}
+/>
+<Select
+	{form}
+	options={model.selectOptions['csf_function']}
+	field="csf_function"
+	label={m.csfFunction()}
+	cacheLock={cacheLocks['csf_function']}
+	bind:cachedValue={formDataCache['csf_function']}
+/>
+<TextArea
+	{form}
+	field="annotation"
+	label={m.annotation()}
+	cacheLock={cacheLocks['annotation']}
+	bind:cachedValue={formDataCache['annotation']}
+/>
+<TextField
+	{form}
+	field="provider"
+	label={m.provider()}
+	cacheLock={cacheLocks['provider']}
+	bind:cachedValue={formDataCache['provider']}
+/>
+<AutocompleteSelect
+	{form}
+	options={getOptions({ objects: model.foreignKeys['folder'] })}
+	field="folder"
+	cacheLock={cacheLocks['folder']}
+	bind:cachedValue={formDataCache['folder']}
+	label={m.domain()}
+	hidden={initialData.folder}
+/>
