+++ frontend/src/lib/components/Forms/ModelForm/RepresentativesForm.svelte
@@ -13,52 +13,50 @@
 	export let data: any = {};
 </script>
 
-{#if model.urlModel === 'representatives'}
-	<TextField
-		{form}
-		field="email"
-		label={m.email()}
-		cacheLock={cacheLocks['email']}
-		bind:cachedValue={formDataCache['email']}
-		data-focusindex="2"
-	/>
-	{#if !data.user}
-		<Checkbox {form} field="create_user" label={m.createUser()} helpText={m.createUserHelpText()} />
-	{/if}
-	<AutocompleteSelect
-		{form}
-		options={getOptions({ objects: model.foreignKeys['entity'] })}
-		field="entity"
-		cacheLock={cacheLocks['entity']}
-		bind:cachedValue={formDataCache['entity']}
-		label={m.entity()}
-	/>
-	<TextField
-		{form}
-		field="first_name"
-		label={m.firstName()}
-		cacheLock={cacheLocks['first_name']}
-		bind:cachedValue={formDataCache['first_name']}
-	/>
-	<TextField
-		{form}
-		field="last_name"
-		label={m.lastName()}
-		cacheLock={cacheLocks['last_name']}
-		bind:cachedValue={formDataCache['last_name']}
-	/>
-	<TextField
-		{form}
-		field="phone"
-		label={m.phone()}
-		cacheLock={cacheLocks['phone']}
-		bind:cachedValue={formDataCache['phone']}
-	/>
-	<TextField
-		{form}
-		field="role"
-		label={m.role()}
-		cacheLock={cacheLocks['role']}
-		bind:cachedValue={formDataCache['role']}
-	/>
+<TextField
+	{form}
+	field="email"
+	label={m.email()}
+	cacheLock={cacheLocks['email']}
+	bind:cachedValue={formDataCache['email']}
+	data-focusindex="2"
+/>
+{#if !data.user}
+	<Checkbox {form} field="create_user" label={m.createUser()} helpText={m.createUserHelpText()} />
 {/if}
+<AutocompleteSelect
+	{form}
+	options={getOptions({ objects: model.foreignKeys['entity'] })}
+	field="entity"
+	cacheLock={cacheLocks['entity']}
+	bind:cachedValue={formDataCache['entity']}
+	label={m.entity()}
+/>
+<TextField
+	{form}
+	field="first_name"
+	label={m.firstName()}
+	cacheLock={cacheLocks['first_name']}
+	bind:cachedValue={formDataCache['first_name']}
+/>
+<TextField
+	{form}
+	field="last_name"
+	label={m.lastName()}
+	cacheLock={cacheLocks['last_name']}
+	bind:cachedValue={formDataCache['last_name']}
+/>
+<TextField
+	{form}
+	field="phone"
+	label={m.phone()}
+	cacheLock={cacheLocks['phone']}
+	bind:cachedValue={formDataCache['phone']}
+/>
+<TextField
+	{form}
+	field="role"
+	label={m.role()}
+	cacheLock={cacheLocks['role']}
+	bind:cachedValue={formDataCache['role']}
+/>
