+++ frontend/src/lib/components/Forms/ModelForm/RiskScenariosForm.svelte
@@ -12,31 +12,29 @@
 	export let initialData: Record<string, any> = {};
 </script>
 
-{#if model.urlModel === 'risk-scenarios'}
-	<AutocompleteSelect
-		{form}
-		options={getOptions({
-			objects: model.foreignKeys['risk_assessment'],
-			label: 'str',
-			extra_fields: [['project', 'str']]
-		})}
-		field="risk_assessment"
-		cacheLock={cacheLocks['risk_assessment']}
-		bind:cachedValue={formDataCache['risk_assessment']}
-		label={m.riskAssessment()}
-		hidden={initialData.risk_assessment}
-	/>
-	<AutocompleteSelect
-		{form}
-		multiple
-		options={getOptions({
-			objects: model.foreignKeys['threats'],
-			extra_fields: [['folder', 'str']],
-			label: 'auto'
-		})}
-		field="threats"
-		cacheLock={cacheLocks['threats']}
-		bind:cachedValue={formDataCache['threats']}
-		label={m.threats()}
-	/>
-{/if}
+<AutocompleteSelect
+	{form}
+	options={getOptions({
+		objects: model.foreignKeys['risk_assessment'],
+		label: 'str',
+		extra_fields: [['project', 'str']]
+	})}
+	field="risk_assessment"
+	cacheLock={cacheLocks['risk_assessment']}
+	bind:cachedValue={formDataCache['risk_assessment']}
+	label={m.riskAssessment()}
+	hidden={initialData.risk_assessment}
+/>
+<AutocompleteSelect
+	{form}
+	multiple
+	options={getOptions({
+		objects: model.foreignKeys['threats'],
+		extra_fields: [['folder', 'str']],
+		label: 'auto'
+	})}
+	field="threats"
+	cacheLock={cacheLocks['threats']}
+	bind:cachedValue={formDataCache['threats']}
+	label={m.threats()}
+/>
